Changed refresh
So when I went back to adjust height it didn't work. I had a realization that it was the multiple refreshes that were most likely causing the problem. So I commented out the two in bookapp.js and just kept the one in the movie.js since I want the list to refresh AFTER all the information has been put in. Now everything works great!